pyproject.toml
[project]
name = "llm-altruism"
version = "0.1.0"
description = "stress-testing LLM alignment through game-theoretic simulation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # core
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    # providers
    "openai>=1.0",
    "anthropic>=0.40",
    "google-genai>=1.0",
    "httpx>=0.27",
    # analysis
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "numpy>=1.26",
    "pandas>=2.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
